Remove explicit rubocop dependency on omnibus
I think this is an issue in Bundler, but let's add a workaround for it.

Currently, if you add a development dependency to a Gemfile through a
gemspec, with the `gemspec` DSL, and later on in the Gemfile you
explicitly add it again, Bundler will silently prefer the latter
specification even if the former is more restrictive.

I think Bundler should either warn, or prefer the most restrictive one,
but to avoid this issue we can remove the duplicated dependency
specification.