Pass `raven_context` to Sentry
So that we get proper differentiation of exceptions according to the
subprocess that failed to run.

Sentry-raven will look into `#raven_context` of the passed exception and
deep merge it into the given context. However, the passed exception here
is `SubprocessFailed` which does not define any `raven_context`. It's
`HelperSubprocessFailed` defining it, so we should pass over that
information.