Enable AWS ECR client to look for creds in default fallback locations (#6622)
This enables the AWS ECR client to look for creds in default fallback
locations if they are not explicitly provided.

NOTE: This will _only_ affect folks running `dependabot-core` by
themselves outside of the Dependabot Service offered by GitHub.

According to the [AWS ECR SDK Client
docs](https://docs.aws.amazon.com/sdk-for-ruby/v3/api/Aws/ECR/Client.html#initialize-instance_method):
> When `:credentials` are not configured directly, the following locations will be searched for credentials:
> * `Aws.config[:credentials]`
> * The `:access_key_id`, `:secret_access_key`, and `:session_token` options.
> * `ENV['AWS_ACCESS_KEY_ID']`, `ENV['AWS_SECRET_ACCESS_KEY']`
> * `~/.aws/credentials`
> * `~/.aws/config`

The flow here of picking what creds to prioritize is important to get correct. We have to cover all the following scenarios:
  1. users on Dependabot Service at GitHub... our proxy will manage the creds, so `dependabot-core` never sees them
  2. users who are hardcoding creds for `dependabot-script`
  3. users who have creds in `~/.aws/`
  4. users accessing public ECR that doesn't require creds

For 1️⃣ , it'll be the same as 4️⃣ , and happens in a context where 2️⃣ / 3️⃣
won't be populated, so should be straightforward as long as there's not
a big perf hit for the ECR SDK library to check the filesystem for 3️⃣.

99% of traffic through this library probably happens through the
Dependabot Service on GitHub, so 99% of the time the checks will be
pointless... but the naive flow is far more intuitive, so if we do need
a flag for perf reasons, we should make it opt-out of credential
checking rather than opt-in. And actually we'd be able to re-use this
across the `dependabot-core` library anywhere we use creds.

Co-authored-by: Jim Cockburn <jim.cockburn@skyscanner.net>