Use the new base image (#7565)
We split the monolith docker image into per-ecosystem images. The
`dependabot/dependabot-core` image was replaced by
`dependabot/dependabot-updater-core`.

Additionally, we don't publish this new image to Dockerhub, but only to
GHCR.

So this corrects the references to it.

This is mostly the great work of `_dwc0011` from https://github.com/dependabot/dependabot-core/pull/7458.

However, the scope of that PR is starting to get a bit larger, and we'll
need to have more conversations about the best way forward. It's unclear
if we should even have a default `FROM` value vs always forcing it to
be set. But that's a bigger conversation.

So for now I extracted this smaller set of changes out so we can at
least stop referencing an old Dockerfile that we no longer update.

Co-authored-by: dwc0011 <53784917+dwc0011@users.noreply.github.com>