Add option to control PR description max length + set reasonable defaults for each platform (#7487)
Add option to set the max character length of the PR description. Any descriptions longer than the max will get truncated with a clear message to the user that the remaining description was truncated. 

This is mostly a refactor / moving of existing code that had been copy/pasted across several PR creators so that we have a clean generic version that can re-used everywhere.

In addition to being configurable, this also sets reasonable defaults for some of the platforms we support.

---------

Co-authored-by: Dennis Carey <dwc0011@users.noreply.github.com>
Co-authored-by: Jeff Widman <jeff@jeffwidman.com>