Match new Git reference not found error message format
The format of this error message changed.

Old message:
```
ERROR:pip.subprocessor:Command errored out with exit status 1:
 command: git checkout -q v15.1.2
      cwd: /tmp/pipenv-4y8m31vy-src/pythonfinder
Complete output (1 lines):
error: pathspec 'v15.1.2' did not match any file(s) known to git
<full traceback>
```

New message:
```
ERROR:pip.subprocessor:[present-rich] git checkout -q v15.1.2 exited with 1
<full traceback>
```

So updated the regex to match the new format.

Unfortunately the dependency name is nowhere to be found in the error
message or full traceback, so we can no longer provide the user with
that helpful information.

I'm planning to reach out to the `pipenv` maintainers to see if they'd
be open to exposing that information again in this error message.