Use defaults instead of comments for documentation
The documentation that was generated when this was installed has been
causing warnings to be added to PR diffs. These values are defaults,
which seem to be just as helpful as the comments which were added.

Source for defaults: https://github.com/Shopify/tapioca/blob/fa61b834119d4a9dc0643862d4ad4804225b0ab9/README.md?plain=1#L879-L935