Expose a new directory_not_found user error
If the user is configuring a directory for dependabot that no longer
exists, and the ecosystem does not clone full repositories, we'll end up
with unknown `Octokit::NotFound` errors when trying to fetch dependency
files.

This commits creates a new user that will be displayed to the user so
that she can fix her configuration.